System.Process.Concurrent

Description

The functions exported by this module are intended to be drop-in replacements for those from System.Process, when converting a whole program to use System.Console.Concurrent.

Documentation

createProcess :: CreateProcess -> IO (Maybe Handle, Maybe Handle, Maybe Handle, ProcessHandle)

Calls createProcessConcurrent

waitForProcess :: ProcessHandle -> IO ExitCode

Calls waitForProcessConcurrent